refactor: update auth server responses, improve form validation, and refine internal component imports
